requirements.txt
django==3.2
django-tailwind==3.0.1
pillow
djangorestframework